+++ exceptions.py
@@ -1,0 +1,11 @@
+class LLMError(Exception):
+    pass
+
+class MemoryError(Exception):
+    pass
+
+class ParserError(Exception):
+    pass
+
+class ToolError(Exception):
+    pass